Refactor cli error checks
