+++ cli/dstack/cli/commands/cp/__init__.py
@@ -8,9 +8,8 @@ from dstack.api.repo import load_repo_data
 from dstack.api.run import RunNotFoundError, TagNotFoundError, get_tagged_run_name
 from dstack.backend.base import Backend
 from dstack.cli.commands import BasicCommand
-from dstack.cli.common import console
+from dstack.cli.common import check_config, check_git, console
 from dstack.core.config import get_dstack_dir
-from dstack.core.error import BackendError, check_config, check_git
 from dstack.core.repo import RepoAddress
