+++ cli/dstack/cli/commands/init/__init__.py
@@ -5,9 +5,8 @@ from typing import Optional
 from dstack.api.backend import list_backends
 from dstack.api.repo import load_repo_data
 from dstack.cli.commands import BasicCommand
-from dstack.cli.common import console
+from dstack.cli.common import check_config, check_git, console
 from dstack.cli.config import BaseConfig
-from dstack.core.error import check_config, check_git
 from dstack.core.userconfig import RepoUserConfig
