+++ cli/dstack/cli/commands/prune/__init__.py
@@ -5,8 +5,7 @@ from dstack.api.backend import list_backends
 from dstack.api.repo import load_repo_data
 from dstack.backend.base import Backend
 from dstack.cli.commands import BasicCommand
-from dstack.cli.common import console
-from dstack.core.error import check_config, check_git
+from dstack.cli.common import check_config, check_git, console
 from dstack.core.repo import LocalRepoData
