+++ cli/dstack/cli/commands/rm/__init__.py
@@ -7,7 +7,7 @@ from rich.prompt import Confirm
 from dstack.api.backend import list_backends
 from dstack.api.repo import load_repo_data
 from dstack.cli.commands import BasicCommand
-from dstack.core.error import check_config, check_git
+from dstack.cli.common import check_config, check_git
 
 
 class RMCommand(BasicCommand):
