+++ cli/dstack/cli/commands/secrets/__init__.py
@@ -10,7 +10,7 @@ from rich_argparse import RichHelpFormatter
 from dstack.api.backend import list_backends
 from dstack.api.repo import load_repo_data
 from dstack.cli.commands import BasicCommand
-from dstack.core.error import check_config, check_git
+from dstack.cli.common import check_config, check_git
 from dstack.core.secret import Secret
