+++ cli/dstack/cli/commands/stop/__init__.py
@@ -5,8 +5,7 @@ from rich.prompt import Confirm
 from dstack.api.backend import list_backends
 from dstack.api.repo import load_repo_data
 from dstack.cli.commands import BasicCommand
-from dstack.cli.common import console
-from dstack.core.error import check_config, check_git
+from dstack.cli.common import check_config, check_git, console
 
 
 def _verb(abort: bool):
