+++ cli/dstack/cli/commands/tags/__init__.py
@@ -12,7 +12,8 @@ from dstack.api.repo import load_repo_data
 from dstack.api.tags import list_tag_heads_with_merged_backends
 from dstack.backend.base import BackendType
 from dstack.cli.commands import BasicCommand
-from dstack.core.error import BackendError, check_config, check_git
+from dstack.cli.common import check_config, check_git
+from dstack.core.error import BackendError
 from dstack.utils.common import pretty_date
