+++ cli/dstack/core/error.py
@@ -1,8 +1,4 @@
-import os
-import sys
 from typing import List, Optional
-
-from git import InvalidGitRepositoryError
 
 
 class ConfigError(Exception):
@@ -25,33 +21,3 @@ class BackendError(Exception):
 class SecretError(Exception):
     def __init__(self, message: Optional[str] = None):
         self.message = message
-
-
-def check_config(func):
-    def decorator(*args, **kwargs):
-        try:
-            func(*args, **kwargs)
-        except ConfigError:
-            sys.exit(f"Call 'dstack config' first")
-
-    return decorator
-
-
-def check_git(func):
-    def decorator(*args, **kwargs):
-        try:
-            func(*args, **kwargs)
-        except InvalidGitRepositoryError:
-            sys.exit(f"{os.getcwd()} is not a Git repo")
-
-    return decorator
-
-
-def check_backend(func):
-    def decorator(*args, **kwargs):
-        try:
-            func(*args, **kwargs)
-        except BackendError as e:
-            sys.exit(e.message)
-
-    return decorator
